+++ CHANGELOG.md
@@ -59,6 +59,7 @@ sharing UI)
 
 ## 0.2 Basic Functionality
 
++ ADD TAG focus to the new TAG text field [0.1.13 20260331 gjw]
 + Add visible last updated [0.1.12 20260331 gjw]
 + Bug - Can not delete Display Name [0.1.11 20260331 gjw]
 + Bug - delete phone retains old labels [0.1.10 20260331 gjw]